Added ElasticsearchConfigBuilder
